+++ Source/Cache/TextureCache.cpp
@@ -18,7 +18,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 NS_DORA_BEGIN
 
 uint64_t Texture2D::_storageSize = 0;
-uint64_t Texture2D::_count = 0;
+uint32_t Texture2D::_count = 0;
 
 Texture2D::Texture2D(bgfx::TextureHandle handle, const bgfx::TextureInfo& info, uint64_t flags)
 	: _handle(handle)
@@ -37,7 +37,7 @@ Texture2D::~Texture2D() {
 	}
 }
 
-uint64_t Texture2D::getCount() {
+uint32_t Texture2D::getCount() {
 	return _count;
 }
